+++ source/agent/com/starsector/prepatcher/agent/PrepatcherAgent.java
@@ -16,7 +16,7 @@ import java.util.Map;
 import java.util.Set;
 
 public final class PrepatcherAgent {
-    public static final String VERSION = "0.17.1";
+    public static final String VERSION = "0.17.2";
     private PrepatcherAgent() {}
 
     public static void premain(String agentArgs, Instrumentation instrumentation) {